requirements.txt
alembic==1.17.2
async-generator==1.10
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Faker==40.1.0
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
selenium==4.40.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.45
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.4
wsproto==1.3.2
WTForms==3.2.1
zipp==3.23.0
